requirements.txt
bs4
google-search-results
python-dotenv
requests
spacy
